+++ servidor/src/controlador/catalogos/catDiscapacidadControlador.ts
@@ -1,10 +1,11 @@
+import { ObtenerRutaLeerPlantillas } from '../../libs/accesoCarpetas';
 import { Request, Response } from 'express';
 import { QueryResult } from 'pg';
+import AUDITORIA_CONTROLADOR from '../auditoria/auditoriaControlador';
 import pool from '../../database';
 import fs from 'fs';
 import path from 'path';
 import excel from 'xlsx';
-import { ObtenerRutaLeerPlantillas } from '../../libs/accesoCarpetas';
 
 class DiscapacidadControlador {
 
@@ -29,7 +30,7 @@ class DiscapacidadControlador {
     // METODO PARA REGISTRAR UN TIPO DE DISCAPACIDAD
     public async CrearDiscapacidad(req: Request, res: Response): Promise<Response> {
         try {
-            const { discapacidad } = req.body;
+            const { discapacidad, user_name, ip } = req.body;
             var VERIFICAR_DISCAPACIDAD = await pool.query(
                 `
                 SELECT * FROM e_cat_discapacidad WHERE UPPER(nombre) = $1
@@ -38,6 +39,10 @@ class DiscapacidadControlador {
 
             if (VERIFICAR_DISCAPACIDAD.rows[0] == undefined || VERIFICAR_DISCAPACIDAD.rows[0] == '') {
                 const discapacidadInsertar = discapacidad.charAt(0).toUpperCase() + discapacidad.slice(1).toLowerCase();
+
+                // INICIAR TRANSACCION
+                await pool.query('BEGIN');
+
                 const response: QueryResult = await pool.query(
                     `
                     INSERT INTO e_cat_discapacidad (nombre) VALUES ($1) RETURNING *
@@ -45,6 +50,21 @@ class DiscapacidadControlador {
                     , [discapacidadInsertar]);
 
                 const [discapacidadInsertada] = response.rows;
+
+
+                // AUDITORIA
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'e_cat_discapacidad',
+                    usuario: user_name,
+                    accion: 'I',
+                    datosOriginales: '',
+                    datosNuevos: JSON.stringify(discapacidadInsertada),
+                    ip,
+                    observacion: null
+                });
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
 
                 if (discapacidadInsertada) {
                     return res.status(200).jsonp({ message: 'Registro guardado.', status: '200' })
@@ -56,6 +76,8 @@ class DiscapacidadControlador {
             }
         }
         catch (error) {
+            // ROLLBACK
+            await pool.query('ROLLBACK');
             return res.status(500).jsonp({ message: 'error', status: '500' });
         }
     }
@@ -63,7 +85,7 @@ class DiscapacidadControlador {
     // METODO PARA EDITAR UN TIPO DE DISCAPACIDAD
     public async EditarDiscapacidad(req: Request, res: Response): Promise<Response> {
         try {
-            const { id, nombre } = req.body;
+            const { id, nombre, user_name, ip } = req.body;
             var VERIFICAR_DISCAPACIDAD = await pool.query(
                 `
                 SELECT * FROM e_cat_discapacidad WHERE UPPER(nombre) = $1 AND NOT id = $2
@@ -72,6 +94,10 @@ class DiscapacidadControlador {
 
             if (VERIFICAR_DISCAPACIDAD.rows[0] == undefined || VERIFICAR_DISCAPACIDAD.rows[0] == '') {
                 const nombreConFormato = nombre.charAt(0).toUpperCase() + nombre.slice(1).toLowerCase();
+
+                // INICIAR TRANSACCION
+                await pool.query('BEGIN');
+
                 const response: QueryResult = await pool.query(
                     `
                     UPDATE e_cat_discapacidad SET nombre = $2
@@ -79,6 +105,21 @@ class DiscapacidadControlador {
                     `
                     , [id, nombreConFormato]);
                 const [discapacidadEditada] = response.rows;
+
+                // AUDITORIA
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'e_cat_discapacidad',
+                    usuario: user_name,
+                    accion: 'U',
+                    datosOriginales: JSON.stringify(VERIFICAR_DISCAPACIDAD.rows),
+                    datosNuevos: JSON.stringify(discapacidadEditada),
+                    ip,
+                    observacion: null
+                });
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+
                 if (discapacidadEditada) {
                     return res.status(200).jsonp({ message: 'Registro actualizado.', status: '200' })
                 } else {
@@ -89,6 +130,8 @@ class DiscapacidadControlador {
             }
         }
         catch (error) {
+            // ROLLBACK
+            await pool.query('ROLLBACK');
             return res.status(500).jsonp({ message: 'error', status: '500' });
         }
     }
@@ -97,14 +140,61 @@ class DiscapacidadControlador {
     public async EliminarRegistro(req: Request, res: Response) {
         try {
             const id = req.params.id;
+            const { user_name, ip } = req.body;
+
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
+
+            //CONSULTAR DATOS DE LA DISCAPACIDAD A ELIMINAR
+            const DISCAPACIDAD = await pool.query(
+                `
+                SELECT * FROM e_cat_discapacidad WHERE id = $1
+                `
+                , [id]);
+            
+            const [datosOriginales] = DISCAPACIDAD.rows;
+
+            if (!datosOriginales) {
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'e_cat_discapacidad',
+                    usuario: user_name,
+                    accion: 'D',
+                    datosOriginales: '',
+                    datosNuevos: '',
+                    ip,
+                    observacion: `Error al eliminar el registro con id: ${id}. Registro no encontrado.`
+                });
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+                return res.status(404).jsonp({ message: 'Registro no encontrado.' });
+            }
+
             await pool.query(
                 `
                 DELETE FROM e_cat_discapacidad WHERE id = $1
                 `
                 , [id]);
+
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'e_cat_discapacidad',
+                usuario: user_name,
+                accion: 'D',
+                datosOriginales: JSON.stringify(datosOriginales),
+                datosNuevos: '',
+                ip,
+                observacion: null
+            });
+
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
+
             res.jsonp({ message: 'Registro eliminado.' });
 
         } catch (error) {
+            // ROLLBACK
+            await pool.query('ROLLBACK');
             return res.jsonp({ message: 'error' });
         }
     }
@@ -244,8 +334,7 @@ class DiscapacidadControlador {
      // REGISTRAR PLANTILLA MODALIDAD_CARGO 
      public async CargarPlantilla(req: Request, res: Response) {
         try {
-            const plantilla = req.body;
-            console.log('datos Discapacidad: ', plantilla);
+            const {plantilla, user_name, ip} = req.body;
             var contador = 1;
             var respuesta: any
 
@@ -253,6 +342,9 @@ class DiscapacidadControlador {
                 // DATOS QUE SE GUARDARAN DE LA PLANTILLA INGRESADA
                 const { item, discapacidad, observacion } = data;
                 const disca = discapacidad.charAt(0).toUpperCase() + discapacidad.slice(1).toLowerCase();
+
+                // INICIAR TRANSACCION
+                await pool.query('BEGIN');
 
                 // REGISTRO DE LOS DATOS DE MODLAIDAD LABORAL
                 const response: QueryResult = await pool.query(
@@ -262,6 +354,20 @@ class DiscapacidadControlador {
                     , [disca]);
 
                 const [discapacidad_emp] = response.rows;
+
+                // AUDITORIA
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'e_cat_discapacidad',
+                    usuario: user_name,
+                    accion: 'I',
+                    datosOriginales: '',
+                    datosNuevos: JSON.stringify(discapacidad_emp),
+                    ip,
+                    observacion: null
+                });
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
 
                 if (contador === plantilla.length) {
                     if (discapacidad_emp) {
@@ -276,6 +382,8 @@ class DiscapacidadControlador {
             });
 
         }catch(error){
+            // ROLLBACK
+            await pool.query('ROLLBACK');
             return res.status(500).jsonp({message: 'Error con el servidor metodo CargarPlantilla', status: '500'});
         }
      }
